requirements.txt
mcp[cli]>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
sse-starlette>=2.1.0
anyio>=4.0.0
httpx>=0.27.0
pylint
pytest
pytest-asyncio